perf(core): encapsulate remove object login in button
