+++ modules/@ergonode/activity-logs/src/locales/en_GB.json
@@ -1,0 +1,7 @@
+{
+    "activityLogs": {
+        "page": {
+            "title": "Users activity logs"
+        }
+    }
+}
