+++ modules/@ergonode/attributes/src/components/Buttons/RemoveAttributeGroupButton.vue
@@ -1,0 +1,88 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+<template>
+    <Button
+        data-cy="delete-attribute-group"
+        :theme="secondaryTheme"
+        :size="smallSize"
+        title="REMOVE GROUP"
+        :disabled="!isAllowedToDelete"
+        @click.native="onRemove">
+        <template #prepend="{ color }">
+            <IconDelete :fill-color="color" />
+        </template>
+    </Button>
+</template>
+
+<script>
+import PRIVILEGES from '@Attributes/config/privileges';
+import {
+    ALERT_TYPE,
+} from '@Core/defaults/alerts';
+import {
+    MODAL_TYPE,
+} from '@Core/defaults/modals';
+import {
+    SIZE,
+    THEME,
+} from '@Core/defaults/theme';
+import Button from '@UI/components/Button/Button';
+import IconDelete from '@UI/components/Icons/Actions/IconDelete';
+import {
+    mapActions,
+} from 'vuex';
+
+export default {
+    name: 'RemoveAttributeButton',
+    components: {
+        Button,
+        IconDelete,
+    },
+    computed: {
+        isAllowedToDelete() {
+            return this.$hasAccess([
+                PRIVILEGES.ATTRIBUTE.delete,
+            ]);
+        },
+        smallSize() {
+            return SIZE.SMALL;
+        },
+        secondaryTheme() {
+            return THEME.SECONDARY;
+        },
+    },
+    methods: {
+        ...mapActions('attributeGroup', [
+            'removeAttributeGroup',
+        ]),
+        onRemove() {
+            this.$confirm({
+                type: MODAL_TYPE.DESTRUCTIVE,
+                title: 'Are you sure you want to delete this group?',
+                applyTitle: 'YES, REMOVE',
+                action: () => this.removeAttributeGroup({
+                    onSuccess: this.onRemoveSuccess,
+                    onError: this.onRemoveError,
+                }),
+            });
+        },
+        onRemoveSuccess() {
+            this.$addAlert({
+                type: ALERT_TYPE.SUCCESS,
+                message: 'Attribute group removed',
+            });
+            this.$router.push({
+                name: 'attribute-groups-grid',
+            });
+        },
+        onRemoveError() {
+            this.$addAlert({
+                type: ALERT_TYPE.ERROR,
+                message: 'Attribute group hasn`t been deleted',
+            });
+        },
+    },
+};
+</script>
