+++ modules/@ergonode/attributes/src/pages/attribute-groups/_group/_id.vue
@@ -3,9 +3,7 @@
  * See LICENSE for license details.
  */
 <template>
-    <AttributeGroupPage
-        :title="code"
-        @remove="onRemove" />
+    <AttributeGroupPage :title="code" />
 </template>
 
 <script>
@@ -13,9 +11,6 @@ import AttributeGroupPage from '@Attributes/components/Pages/AttributeGroupPage'
 import {
     ALERT_TYPE,
 } from '@Core/defaults/alerts';
-import {
-    MODAL_TYPE,
-} from '@Core/defaults/modals';
 import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
 import {
     mapActions,
@@ -64,7 +59,6 @@ export default {
     },
     methods: {
         ...mapActions('attributeGroup', [
-            'removeAttributeGroup',
             '__clearStorage',
         ]),
         ...mapActions('feedback', {
@@ -73,32 +67,6 @@ export default {
         ...mapActions('tab', {
             __clearTranslationsStorage: '__clearStorage',
         }),
-        onRemove() {
-            this.$confirm({
-                type: MODAL_TYPE.DESTRUCTIVE,
-                title: 'Are you sure you want to delete this group?',
-                applyTitle: 'YES, REMOVE',
-                action: () => this.removeAttributeGroup({
-                    onSuccess: this.onRemoveSuccess,
-                    onError: this.onRemoveError,
-                }),
-            });
-        },
-        onRemoveSuccess() {
-            this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Attribute group removed',
-            });
-            this.$router.push({
-                name: 'attribute-groups-grid',
-            });
-        },
-        onRemoveError() {
-            this.$addAlert({
-                type: ALERT_TYPE.ERROR,
-                message: 'Attribute group hasn`t been deleted',
-            });
-        },
     },
     head() {
         return {
