+++ modules/@ergonode/collections/src/pages/collections/_collection/_id.vue
@@ -3,9 +3,7 @@
  * See LICENSE for license details.
  */
 <template>
-    <CollectionPage
-        :title="code"
-        @remove="onRemove" />
+    <CollectionPage :title="code" />
 </template>
 
 <script>
@@ -13,9 +11,6 @@ import CollectionPage from '@Collections/components/Pages/CollectionPage';
 import {
     ALERT_TYPE,
 } from '@Core/defaults/alerts';
-import {
-    MODAL_TYPE,
-} from '@Core/defaults/modals';
 import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
 import {
     mapActions,
@@ -64,7 +59,6 @@ export default {
     },
     methods: {
         ...mapActions('collection', [
-            'removeCollection',
             '__clearStorage',
         ]),
         ...mapActions('feedback', {
@@ -73,32 +67,6 @@ export default {
         ...mapActions('tab', {
             __clearTranslationsStorage: '__clearStorage',
         }),
-        onRemove() {
-            this.$confirm({
-                type: MODAL_TYPE.DESTRUCTIVE,
-                title: 'Are you sure you want to delete this collection?',
-                applyTitle: 'YES, REMOVE',
-                action: () => this.removeCollection({
-                    onSuccess: this.onRemoveSuccess,
-                    onError: this.onRemoveError,
-                }),
-            });
-        },
-        onRemoveSuccess() {
-            this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Product collection removed',
-            });
-            this.$router.push({
-                name: 'collections-grid',
-            });
-        },
-        onRemoveError() {
-            this.$addAlert({
-                type: ALERT_TYPE.ERROR,
-                message: 'Collection hasn`t been deleted',
-            });
-        },
     },
     head() {
         return {
