+++ modules/@ergonode/core/src/components/Buttons/RemoveUnitButton.vue
@@ -1,0 +1,94 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+<template>
+    <Button
+        data-cy="delete-unit"
+        :theme="secondaryTheme"
+        :size="smallSize"
+        title="REMOVE UNIT"
+        :disabled="!isAllowedToDelete"
+        @click.native="onRemove">
+        <template #prepend="{ color }">
+            <IconDelete :fill-color="color" />
+        </template>
+    </Button>
+</template>
+
+<script>
+import PRIVILEGES from '@Core/config/privileges';
+import {
+    ALERT_TYPE,
+} from '@Core/defaults/alerts';
+import {
+    MODAL_TYPE,
+} from '@Core/defaults/modals';
+import {
+    SIZE,
+    THEME,
+} from '@Core/defaults/theme';
+import Button from '@UI/components/Button/Button';
+import IconDelete from '@UI/components/Icons/Actions/IconDelete';
+import {
+    mapActions,
+} from 'vuex';
+
+export default {
+    name: 'RemoveUnitButton',
+    components: {
+        Button,
+        IconDelete,
+    },
+    computed: {
+        isAllowedToDelete() {
+            return this.$hasAccess([
+                PRIVILEGES.SETTINGS.delete,
+            ]);
+        },
+        smallSize() {
+            return SIZE.SMALL;
+        },
+        secondaryTheme() {
+            return THEME.SECONDARY;
+        },
+    },
+    methods: {
+        ...mapActions('unit', [
+            'removeUnit',
+        ]),
+        ...mapActions('dictionaries', [
+            'getDictionary',
+        ]),
+        onRemove() {
+            this.$confirm({
+                type: MODAL_TYPE.DESTRUCTIVE,
+                title: 'Are you sure you want to delete this unit?',
+                applyTitle: 'YES, REMOVE',
+                action: () => this.removeUnit({
+                    onSuccess: this.onRemoveSuccess,
+                    onError: this.onRemoveError,
+                }),
+            });
+        },
+        async onRemoveSuccess() {
+            await this.getDictionary({
+                dictionaryName: 'units',
+            });
+            await this.$addAlert({
+                type: ALERT_TYPE.SUCCESS,
+                message: 'Unit removed',
+            });
+            await this.$router.push({
+                name: 'settings-units',
+            });
+        },
+        onRemoveError() {
+            this.$addAlert({
+                type: ALERT_TYPE.ERROR,
+                message: 'Unit hasn`t been deleted',
+            });
+        },
+    },
+};
+</script>
