+++ modules/@ergonode/core/src/pages/settings/_unit/_id.vue
@@ -3,9 +3,7 @@
  * See LICENSE for license details.
  */
 <template>
-    <UnitPage
-        :title="name"
-        @remove="onRemove" />
+    <UnitPage :title="name" />
 </template>
 
 <script>
@@ -13,9 +11,6 @@ import UnitPage from '@Core/components/Pages/UnitPage';
 import {
     ALERT_TYPE,
 } from '@Core/defaults/alerts';
-import {
-    MODAL_TYPE,
-} from '@Core/defaults/modals';
 import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
 import {
     mapActions,
@@ -62,41 +57,8 @@ export default {
     },
     methods: {
         ...mapActions('unit', {
-            removeUnit: 'removeUnit',
             clearUnitStorage: '__clearStorage',
         }),
-        ...mapActions('dictionaries', [
-            'getDictionary',
-        ]),
-        onRemove() {
-            this.$confirm({
-                type: MODAL_TYPE.DESTRUCTIVE,
-                title: 'Are you sure you want to delete this unit?',
-                applyTitle: 'YES, REMOVE',
-                action: () => this.removeUnit({
-                    onSuccess: this.onRemoveSuccess,
-                    onError: this.onRemoveError,
-                }),
-            });
-        },
-        async onRemoveSuccess() {
-            await this.getDictionary({
-                dictionaryName: 'units',
-            });
-            await this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Unit removed',
-            });
-            await this.$router.push({
-                name: 'settings-units',
-            });
-        },
-        onRemoveError() {
-            this.$addAlert({
-                type: ALERT_TYPE.ERROR,
-                message: 'Unit hasn`t been deleted',
-            });
-        },
     },
     head() {
         return {
