+++ modules/@ergonode/media/src/components/Buttons/RemoveResourceButton.vue
@@ -1,0 +1,87 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+<template>
+    <Button
+        :theme="secondaryTheme"
+        :size="smallSize"
+        title="REMOVE RESOURCE"
+        :disabled="!isAllowedToDelete"
+        @click.native="onRemove">
+        <template #prepend="{ color }">
+            <IconDelete :fill-color="color" />
+        </template>
+    </Button>
+</template>
+
+<script>
+import {
+    ALERT_TYPE,
+} from '@Core/defaults/alerts';
+import {
+    MODAL_TYPE,
+} from '@Core/defaults/modals';
+import {
+    SIZE,
+    THEME,
+} from '@Core/defaults/theme';
+import PRIVILEGES from '@Media/config/privileges';
+import Button from '@UI/components/Button/Button';
+import IconDelete from '@UI/components/Icons/Actions/IconDelete';
+import {
+    mapActions,
+} from 'vuex';
+
+export default {
+    name: 'RemoveResourceButton',
+    components: {
+        Button,
+        IconDelete,
+    },
+    computed: {
+        isAllowedToDelete() {
+            return this.$hasAccess([
+                PRIVILEGES.MULTIMEDIA.delete,
+            ]);
+        },
+        smallSize() {
+            return SIZE.SMALL;
+        },
+        secondaryTheme() {
+            return THEME.SECONDARY;
+        },
+    },
+    methods: {
+        ...mapActions('media', [
+            'removeResource',
+        ]),
+        onRemove() {
+            this.$confirm({
+                type: MODAL_TYPE.DESTRUCTIVE,
+                title: 'Are you sure you want to delete this resource?',
+                applyTitle: 'YES, REMOVE',
+                action: () => this.removeResource({
+                    onSuccess: this.onRemoveSuccess,
+                    onError: this.onRemoveError,
+                }),
+            });
+        },
+        onRemoveSuccess() {
+            this.$addAlert({
+                type: ALERT_TYPE.SUCCESS,
+                message: 'Resource removed',
+            });
+            this.$router.push({
+                name: 'media-grid',
+            });
+        },
+        onRemoveError() {
+            this.$addAlert({
+                type: ALERT_TYPE.ERROR,
+                message: 'Resource hasn`t been deleted',
+            });
+        },
+    },
+};
+</script>
