+++ modules/@ergonode/product-templates/src/components/Buttons/RemoveProductTemplateButton.vue
@@ -1,0 +1,85 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+<template>
+    <Button
+        data-cy="delete-template"
+        :theme="secondaryTheme"
+        :size="smallSize"
+        title="REMOVE TEMPLATE"
+        :disabled="!isAllowedToDelete"
+        @click.native="onRemove">
+        <template #prepend="{ color }">
+            <IconDelete
+                :fill-color="color" />
+        </template>
+    </Button>
+</template>
+
+<script>
+import {
+    ALERT_TYPE,
+} from '@Core/defaults/alerts';
+import {
+    MODAL_TYPE,
+} from '@Core/defaults/modals';
+import {
+    SIZE,
+    THEME,
+} from '@Core/defaults/theme';
+import PRIVILEGES from '@Templates/config/privileges';
+import Button from '@UI/components/Button/Button';
+import IconDelete from '@UI/components/Icons/Actions/IconDelete';
+import {
+    mapActions,
+} from 'vuex';
+
+export default {
+    name: 'RemoveProductTemplateButton',
+    components: {
+        Button,
+        IconDelete,
+    },
+    computed: {
+        isAllowedToDelete() {
+            return this.$hasAccess([
+                PRIVILEGES.TEMPLATE_DESIGNER.delete,
+            ]);
+        },
+        smallSize() {
+            return SIZE.SMALL;
+        },
+        secondaryTheme() {
+            return THEME.SECONDARY;
+        },
+    },
+    methods: {
+        ...mapActions('productTemplate', [
+            'removeTemplate',
+        ]),
+        onRemoveSuccess() {
+            this.$addAlert({
+                type: ALERT_TYPE.SUCCESS,
+                message: 'Template removed',
+            });
+            this.$router.push({
+                name: 'product-templates',
+            });
+        },
+        onRemove() {
+            this.$confirm({
+                type: MODAL_TYPE.DESTRUCTIVE,
+                title: 'Are you sure you want to delete this template?',
+                applyTitle: 'YES, REMOVE',
+                action: () => {
+                    this.removeTemplate({
+                        scope: this.scope,
+                        onSuccess: this.onRemoveSuccess,
+                    });
+                },
+            });
+        },
+    },
+};
+</script>
