+++ modules/@ergonode/product-templates/src/components/Pages/TemplatePage.vue
@@ -11,18 +11,7 @@
                 <NavigationBackFab />
             </template>
             <template #mainAction>
-                <Button
-                    data-cy="delete-template"
-                    :theme="secondaryTheme"
-                    :size="smallSize"
-                    title="REMOVE TEMPLATE"
-                    :disabled="!isAllowedToDelete"
-                    @click.native="onRemove">
-                    <template #prepend="{ color }">
-                        <IconDelete
-                            :fill-color="color" />
-                    </template>
-                </Button>
+                <RemoveProductTemplateButton />
             </template>
         </TitleBar>
         <HorizontalRoutingTabBar
@@ -36,10 +25,14 @@
 <script>
 import editPageMixin from '@Core/mixins/page/editPageMixin';
 import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
+import RemoveProductTemplateButton from '@Templates/components/Buttons/RemoveProductTemplateButton';
 import PRIVILEGES from '@Templates/config/privileges';
 
 export default {
     name: 'TemplatePage',
+    components: {
+        RemoveProductTemplateButton,
+    },
     mixins: [
         editPageMixin,
         asyncTabsMixin,
@@ -47,11 +40,6 @@ export default {
     computed: {
         isReadOnly() {
             return this.$isReadOnly(PRIVILEGES.TEMPLATE_DESIGNER.namespace);
-        },
-        isAllowedToDelete() {
-            return this.$hasAccess([
-                PRIVILEGES.TEMPLATE_DESIGNER.delete,
-            ]);
         },
     },
 };
