+++ modules/@ergonode/segments/src/pages/segments/_segment/_id.vue
@@ -3,18 +3,13 @@
  * See LICENSE for license details.
  */
 <template>
-    <SegmentPage
-        :title="code"
-        @remove="onRemove" />
+    <SegmentPage :title="code" />
 </template>
 
 <script>
 import {
     ALERT_TYPE,
 } from '@Core/defaults/alerts';
-import {
-    MODAL_TYPE,
-} from '@Core/defaults/modals';
 import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
 import SegmentPage from '@Segments/components/Pages/SegmentPage';
 import {
@@ -68,7 +63,6 @@ export default {
             __clearConditionStorage: '__clearStorage',
         }),
         ...mapActions('segment', {
-            removeSegment: 'removeSegment',
             __clearSegmentStorage: '__clearStorage',
         }),
         ...mapActions('feedback', {
@@ -77,32 +71,6 @@ export default {
         ...mapActions('gridDesigner', {
             __clearGridDesignerStorage: '__clearStorage',
         }),
-        onRemove() {
-            this.$confirm({
-                type: MODAL_TYPE.DESTRUCTIVE,
-                title: 'Are you sure you want to delete this segment?',
-                applyTitle: 'YES, REMOVE',
-                action: () => this.removeSegment({
-                    onSuccess: this.onRemoveSuccess,
-                    onError: this.onRemoveError,
-                }),
-            });
-        },
-        onRemoveSuccess() {
-            this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Segment removed',
-            });
-            this.$router.push({
-                name: 'segments-grid',
-            });
-        },
-        onRemoveError() {
-            this.$addAlert({
-                type: ALERT_TYPE.ERROR,
-                message: 'Segmet hasn`t been deleted',
-            });
-        },
     },
     head() {
         return {
