+++ modules/@ergonode/status-transitions/src/components/Pages/TransitionPage.vue
@@ -11,16 +11,7 @@
                 <NavigationBackFab />
             </template>
             <template #mainAction>
-                <Button
-                    :theme="secondaryTheme"
-                    :size="smallSize"
-                    title="REMOVE TRANSITION"
-                    :disabled="!isAllowedToDelete"
-                    @click.native="onRemove">
-                    <template #prepend="{ color }">
-                        <IconDelete :fill-color="color" />
-                    </template>
-                </Button>
+                <RemoveStatusTransitionButton />
             </template>
         </TitleBar>
         <HorizontalRoutingTabBar
@@ -34,10 +25,14 @@
 <script>
 import editPageMixin from '@Core/mixins/page/editPageMixin';
 import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
+import RemoveStatusTransitionButton from '@Transitions/components/Buttons/RemoveStatusTransitionButton';
 import PRIVILEGES from '@Transitions/config/privileges';
 
 export default {
     name: 'TransitionPage',
+    components: {
+        RemoveStatusTransitionButton,
+    },
     mixins: [
         editPageMixin,
         asyncTabsMixin,
@@ -45,11 +40,6 @@ export default {
     computed: {
         isReadOnly() {
             return this.$isReadOnly(PRIVILEGES.WORKFLOW.namespace);
-        },
-        isAllowedToDelete() {
-            return this.$hasAccess([
-                PRIVILEGES.WORKFLOW.delete,
-            ]);
         },
     },
 };
