+++ modules/@ergonode/status-transitions/src/pages/status-transitions/_transition/_id.vue
@@ -3,18 +3,13 @@
  * See LICENSE for license details.
  */
 <template>
-    <TransitionPage
-        :title="`${source.value} -> ${destination.value}`"
-        @remove="onRemove" />
+    <TransitionPage :title="`${source.value} -> ${destination.value}`" />
 </template>
 
 <script>
 import {
     ALERT_TYPE,
 } from '@Core/defaults/alerts';
-import {
-    MODAL_TYPE,
-} from '@Core/defaults/modals';
 import TransitionPage from '@Transitions/components/Pages/TransitionPage';
 import {
     mapActions,
@@ -66,38 +61,11 @@ export default {
             __clearConditionStorage: '__clearStorage',
         }),
         ...mapActions('statusTransition', {
-            removeStatusTransition: 'removeStatusTransition',
             __clearTransitionStorage: '__clearStorage',
         }),
         ...mapActions('feedback', {
             __clearFeedbackStorage: '__clearStorage',
         }),
-        onRemove() {
-            this.$confirm({
-                type: MODAL_TYPE.DESTRUCTIVE,
-                title: 'Are you sure you want to delete this transition?',
-                applyTitle: 'YES, REMOVE',
-                action: () => this.removeStatusTransition({
-                    onSuccess: this.onRemoveSuccess,
-                    onError: this.onRemoveError,
-                }),
-            });
-        },
-        onRemoveSuccess() {
-            this.$addAlert({
-                type: ALERT_TYPE.SUCCESS,
-                message: 'Transition removed',
-            });
-            this.$router.push({
-                name: 'status-transitions-grid',
-            });
-        },
-        onRemoveError() {
-            this.$addAlert({
-                type: ALERT_TYPE.ERROR,
-                message: 'Transition hasn`t been deleted',
-            });
-        },
     },
     head() {
         return {
