+++ modules/@ergonode/users/src/components/Buttons/RemoveRoleButton.vue
@@ -1,0 +1,87 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+<template>
+    <Button
+        :theme="secondaryTheme"
+        :size="smallSize"
+        title="REMOVE ROLE"
+        :disabled="!isAllowedToDelete"
+        @click.native="onRemove">
+        <template #prepend="{ color }">
+            <IconDelete :fill-color="color" />
+        </template>
+    </Button>
+</template>
+
+<script>
+import {
+    ALERT_TYPE,
+} from '@Core/defaults/alerts';
+import {
+    MODAL_TYPE,
+} from '@Core/defaults/modals';
+import {
+    SIZE,
+    THEME,
+} from '@Core/defaults/theme';
+import Button from '@UI/components/Button/Button';
+import IconDelete from '@UI/components/Icons/Actions/IconDelete';
+import PRIVILEGES from '@Users/config/privileges';
+import {
+    mapActions,
+} from 'vuex';
+
+export default {
+    name: 'RemoveRoleButton',
+    components: {
+        Button,
+        IconDelete,
+    },
+    computed: {
+        isAllowedToDelete() {
+            return this.$hasAccess([
+                PRIVILEGES.USER_ROLE.delete,
+            ]);
+        },
+        smallSize() {
+            return SIZE.SMALL;
+        },
+        secondaryTheme() {
+            return THEME.SECONDARY;
+        },
+    },
+    methods: {
+        ...mapActions('role', [
+            'removeRole',
+        ]),
+        onRemove() {
+            this.$confirm({
+                type: MODAL_TYPE.DESTRUCTIVE,
+                title: 'Are you sure you want to delete this role?',
+                applyTitle: 'YES, REMOVE',
+                action: () => this.removeRole({
+                    onSuccess: this.onRemoveSuccess,
+                    onError: this.onRemoveError,
+                }),
+            });
+        },
+        onRemoveSuccess() {
+            this.$addAlert({
+                type: ALERT_TYPE.SUCCESS,
+                message: 'Role removed',
+            });
+            this.$router.push({
+                name: 'user-roles-grid',
+            });
+        },
+        onRemoveError() {
+            this.$addAlert({
+                type: ALERT_TYPE.ERROR,
+                message: 'Role hasn`t been deleted',
+            });
+        },
+    },
+};
+</script>
